Bugfix for mongoose not allowing nested schema
Mongoose requires that nested schema either be a reference to the other
document or an array of the schema.
Here you will see a dirty hack, as I nest them. The array length should
ALWAYS be one.
Why am I doing this? For code maintainability on the cart schema side.
Also because we delete the cart after the order is made from the carts
collection.